Detach 214 notes, overrides
