+++ gvm/protocols/gmpv214/entities/notes.py
@@ -1,0 +1,177 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2021 Greenbone Networks GmbH
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+# pylint:  disable=redefined-builtin
+# MAYBE we should change filter to filter_string (everywhere)
+
+from typing import Any, List, Optional
+
+from gvm.errors import RequiredArgument
+from gvm.protocols.gmpv208.entities.notes import NotesMixin as Gmp208NotesMixin
+from gvm.protocols.gmpv208.entities.severity import Severity, SeverityLevel
+from gvm.utils import deprecation, to_comma_list
+from gvm.xml import XmlCommand
+
+
+class NotesMixin(Gmp208NotesMixin):
+    def create_note(
+        self,
+        text: str,
+        nvt_oid: str,
+        *,
+        days_active: Optional[int] = None,
+        hosts: Optional[List[str]] = None,
+        port: Optional[int] = None,
+        result_id: Optional[str] = None,
+        severity: Optional[Severity] = None,
+        task_id: Optional[str] = None,
+        threat: Optional[SeverityLevel] = None,
+    ) -> Any:
+        """Create a new note
+
+        Arguments:
+            text: Text of the new note
+            nvt_id: OID of the nvt to which note applies
+            days_active: Days note will be active. -1 on
+                always, 0 off
+            hosts: A list of hosts addresses
+            port: Port to which the note applies
+            result_id: UUID of a result to which note applies
+            severity: Severity to which note applies
+            task_id: UUID of task to which note applies
+            threat: Severity level to which note applies. Will be converted to
+                severity.
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not text:
+            raise RequiredArgument(
+                function=self.create_note.__name__, argument='text'
+            )
+
+        if not nvt_oid:
+            raise RequiredArgument(
+                function=self.create_note.__name__, argument='nvt_oid'
+            )
+
+        cmd = XmlCommand("create_note")
+        cmd.add_element("text", text)
+        cmd.add_element("nvt", attrs={"oid": nvt_oid})
+
+        if days_active is not None:
+            cmd.add_element("active", str(days_active))
+
+        if hosts:
+            cmd.add_element("hosts", to_comma_list(hosts))
+
+        if port:
+            cmd.add_element("port", str(port))
+
+        if result_id:
+            cmd.add_element("result", attrs={"id": result_id})
+
+        if severity:
+            cmd.add_element("severity", str(severity))
+
+        if task_id:
+            cmd.add_element("task", attrs={"id": task_id})
+
+        if threat is not None:
+            deprecation(
+                "The threat parameter has been removed in GMP"
+                " version {}{}".format(
+                    self.get_protocol_version()[0],
+                    self.get_protocol_version()[1],
+                )
+            )
+
+        return self._send_xml_command(cmd)
+
+    def modify_note(
+        self,
+        note_id: str,
+        text: str,
+        *,
+        days_active: Optional[int] = None,
+        hosts: Optional[List[str]] = None,
+        port: Optional[int] = None,
+        result_id: Optional[str] = None,
+        severity: Optional[Severity] = None,
+        task_id: Optional[str] = None,
+        threat: Optional[SeverityLevel] = None,
+    ) -> Any:
+        """Modifies an existing note.
+
+        Arguments:
+            note_id: UUID of note to modify.
+            text: The text of the note.
+            days_active: Days note will be active. -1 on always, 0 off.
+            hosts: A list of hosts addresses
+            port: Port to which note applies.
+            result_id: Result to which note applies.
+            severity: Severity to which note applies.
+            task_id: Task to which note applies.
+            threat: Threat level to which note applies. Will be converted to
+                severity.
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not note_id:
+            raise RequiredArgument(
+                function=self.modify_note.__name__, argument='note_id'
+            )
+
+        if not text:
+            raise RequiredArgument(
+                function=self.modify_note.__name__, argument='text'
+            )
+
+        cmd = XmlCommand("modify_note")
+        cmd.set_attribute("note_id", note_id)
+        cmd.add_element("text", text)
+
+        if days_active is not None:
+            cmd.add_element("active", str(days_active))
+
+        if hosts:
+            cmd.add_element("hosts", to_comma_list(hosts))
+
+        if port:
+            cmd.add_element("port", str(port))
+
+        if result_id:
+            cmd.add_element("result", attrs={"id": result_id})
+
+        if severity:
+            cmd.add_element("severity", str(severity))
+
+        if task_id:
+            cmd.add_element("task", attrs={"id": task_id})
+
+        if threat is not None:
+            deprecation(
+                "The threat parameter has been removed in GMP"
+                " version {}{}".format(
+                    self.get_protocol_version()[0],
+                    self.get_protocol_version()[1],
+                )
+            )
+
+        return self._send_xml_command(cmd)
